Install The NERD tree.
The NERD tree is a tree explorer plugin for navigating the filesystem.
The NERD tree allows you to explore your filesystem and to open files
and directories. It presents the filesystem to you in the form of a tree
which you manipulate with the keyboard and/or mouse. It also allows you
to perform simple filesystem operations.

See more at http://www.vim.org/scripts/script.php?script_id=1658